Complete overhaul!
Inspired by @holman, I've decided to organize all my dotfiles by topic.
I'm using his Rakefile, which sets up symlinks in $HOME that point to
the corresponding files in the repo (append '.symlink' to a filename to
have the Rakefile link it).

Also, I'm using zsh with the agnoster theme, which is dope.